% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR.R
\name{FDR}
\alias{FDR}
\title{FDR estimation for given z-values / p-values}
\usage{
FDR(
  z,
  tol = 5e-06,
  p_value = FALSE,
  local = FALSE,
  alternative = "greater",
  max_iter = 30,
  mono = TRUE,
  thre.z = 0.9,
  Uthre.gam = 0.9,
  Lthre.gam = 0.01
)
}
\arguments{
\item{z}{Matrix which column indicates z-values, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm. If maximum absolute difference
of current and previous gamma value is smaller than tol,
i.e. \eqn{max_i |\gamma_i^{(k+1)}-\gamma_i^{(k)} <tol}, for k-th step,
then optimization stops. (default: 5e-6)}

\item{p_value}{If TRUE, input are p-values. If FALSE, input are z-values. (default: FALSE)}

\item{local}{If TRUE, \code{FDR} returns localFDR estimates for given z-values or p-values. IF FALSE, \code{FDR} returns FDR estimates. (default: FALSE)}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided", "greater" (default) or "less". You can specify just the initial letter. (default: "greater")}

\item{max_iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{mono}{If TRUE, FDR is in ascending order of z-values. (default: TRUE)}

\item{thre_z}{Threshold value which only z-values smaller than thre.z
are used to compute the log-concave estimates f_1 in M-step.}

\item{Uthre_gam}{Upper threshold of gamma which are used to compute stopping criteria for the EM algorithm.}

\item{Lthre_gam}{Lower threshold of gamma which are used to compute stopping criteria for the EM algorithm.}
}
\value{
Estimates of FDR or localFDR for given z-values / p-values.

  \item{FDR}{FDR estimates for given z-values / p-values}
}
\description{
\code{FDR} returns FDR estimates for given multi-dimensional lists of z-values / p-values.
\code{FDR} imports \code{SpMix} for a two-component semiparametric
mixture model to estimate the FDR from the z-values / p-values.
}
