% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFDR.R
\name{localFDR}
\alias{localFDR}
\title{Local False Discovery Rate (localFDR) Estimation}
\usage{
localFDR(
  z,
  tol = 5e-06,
  is_pvalue = FALSE,
  alternative = NULL,
  max_iter = 30,
  monotone = TRUE,
  Uthre_gam = 0.99,
  Lthre_gam = 0.01
)
}
\arguments{
\item{z}{A numeric matrix where each row represents a data point (z-values, p-values, or raw data).}

\item{tol}{Convergence threshold for the EM algorithm. The optimization stops when 
the maximum absolute difference between consecutive gamma values is below `tol`. (default: 5e-6)}

\item{is_pvalue}{Logical; if `TRUE`, the input is assumed to be p-values and transformed accordingly. (default: `FALSE`)}

\item{alternative}{A logical vector specifying whether the alternative distribution is greater (`TRUE`) 
or less (`FALSE`) than the null distribution in each dimension. Must match the number of columns in `z`. 
(default: `NULL`, assumes `greater` for all dimensions)}

\item{max_iter}{Maximum number of iterations for the EM algorithm. (default: 30)}

\item{monotone}{Logical; if `TRUE`, ensures localFDR values are non-decreasing along the z-values. (default: `TRUE`)}

\item{Uthre_gam}{Upper threshold for gamma to determine stopping criteria in the EM algorithm. (default: 0.99)}

\item{Lthre_gam}{Lower threshold for gamma to determine stopping criteria in the EM algorithm. (default: 0.01)}
}
\value{
A numeric vector of estimated local false discovery rates (localFDR).
}
\description{
`localFDR` computes local false discovery rate (localFDR) estimates for multi-dimensional input data 
(raw data, z-values, or p-values). It uses the semiparametric mixture model from `SpMix()`, 
which integrates Efron's empirical null principle and log-concave density estimation.
}
