% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microarrays}
\alias{microarrays}
\title{P-values for Microarray Data}
\format{
A data frame with three variables: \code{expm1}, \code{expm2}, and
\code{expm3}.
}
\usage{
microarrays
}
\description{
List of p-values for three different experiments.
}
\keyword{datasets}
