% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{P-values for Pathways.}
\format{
A data frame with four variables: \code{Gene.Set}, \code{L}
(peripheral leukocytes), \code{O} (orbital inflammatory), and
\code{S} (sinus brushings).
}
\usage{
pathways
}
\description{
The list of significantly upregulated genes in three published studies:
(1) peripheral leukocytes(L), (2) orbital inflammatory disease(O), and
(3) sinus brushings(S) compared to healthy controls.
}
\keyword{datasets}
