% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpMix.R
\name{plot.SpMix}
\alias{plot.SpMix}
\title{Visualization of Fitted SpMix Model}
\usage{
\method{plot}{SpMix}(
  x,
  fdr_cutoff = 0.2,
  testing = TRUE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  coord_legend = c(8, -5, 0.2)
)
}
\arguments{
\item{x}{An object of class "SpMix", obtained from `SpMix()`.}

\item{fdr_cutoff}{Threshold of local false discovery rate (localFDR) for highlighting significant data points (default: 0.2).}

\item{testing}{Logical; if `TRUE`, plots for hypothesis testing with null and alternative distributions. If `FALSE`, plots density estimation (default: `TRUE`).}

\item{xlab}{Label for the x-axis (default: "x").}

\item{ylab}{Label for the y-axis in 2D/3D plots (default: "y").}

\item{zlab}{Label for the z-axis in 3D plots (default: "z").}

\item{coord_legend}{Coordinates for the legend in a 3D scatter plot when data is 2D or 3D (default: `c(8, -5, 0.2)`).}
}
\value{
A visualization of the fitted `SpMix` model. Additionally, it returns:

  - `thre`: The threshold z-value for distinguishing between null and alternative distributions.
}
\description{
`plot.SpMix` visualizes the fitted semiparametric mixture model for objects of class "SpMix". 
It provides:
- A **histogram** with fitted densities for univariate data (1D)
- A **scatter plot** for 2D or 3D data, customizable with the `ggplot2` package.
- When `testing = TRUE` (default), it overlays null and alternative fitted densities for hypothesis testing.
}
