% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mix.1D.R
\name{sp.mix.1D}
\alias{sp.mix.1D}
\title{LocalFDR estimation for 1-dimensional z-values}
\usage{
sp.mix.1D(
  z,
  tol = 5e-06,
  max.iter = 30,
  doplot = TRUE,
  thre.localFDR = 0.2,
  thre.z = 0.95,
  Uthre.gam = 0.9,
  Lthre.gam = 0.01
)
}
\arguments{
\item{z}{Vector which each element indicates z-values, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm. If maximum absolute difference
of current and previous gamma value is smaller than tol,
i.e. \eqn{max_i |\gamma_i^{(k+1)}-\gamma_i^{(k)} <tol}, for k-th step, then optimization stops. (default: 5e-6)}

\item{max.iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{doplot}{Boolean parameter that draws histogram and fitted lines for estimated density,
if TRUE. (default: TRUE)}

\item{thre.localFDR}{Threshold of LocalFDR which is used for calculation of
false positive rate (FPR) and sensitivity with the threshold set to be
\eqn{fdr \leq thre.localFDR}. (default: 0.2)}

\item{thre.z}{Threshold value which only z-values smaller than thre.z
are used to compute the log-concave estimates f_1 in M-step.}

\item{Uthre.gam}{Upper threshold of gamma which are used to compute stopping criteria for the EM algorithm.}

\item{Lthre.gam}{Lower threshold of gamma which are used to compute stopping criteria for the EM algorithm.}
}
\value{
Estimates of semiparametric mixture model for f for given z-values.

  \item{p.0}{Prior probability for null distribution}
  \item{mu.0 sig.0}{Parameter estimates of normal null distribution, N(mu.0, sig.0^2)}
  \item{f}{Probability estimates of semiparametric mixture model for each z-value point.}
  \item{localfdr}{LocalFDR estimates for given z-values}
  \item{iter}{Number of iterations of EM algorithm to compute LocalFDR.}
}
\description{
\code{sp.mix.1D} returns LocalFDR estimates and semiparametric
mixture density estimates for given 1-dimensional lists of z-values, which
are the probit-transformed p-values.
For the hypothesis testing \code{sp.mix.1D} uses a two-component semiparametric
mixture model to estimate the LocalFDR from the p-values. The two pillars of
the proposed approach are Efron's empirical null principle and log-concave
density estimation for the alternative distribution.
}
