% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPMix.R
\name{SPMix}
\alias{SPMix}
\title{Semiparametric Mixture Model for Local False Discovery Rate Estimation}
\usage{
SPMix(
  z,
  initial_p0 = 0.5,
  tol = 5e-05,
  is_pvalue = FALSE,
  alternative = NULL,
  min_iter = 3,
  max_iter = 50,
  thre_z = 0.5,
  Uthre_gam = 0.99,
  Lthre_gam = 0.01,
  thre = 0.2
)
}
\arguments{
\item{z}{A numeric matrix where each row represents a data point (z-values, p-values, or raw data).}

\item{initial_p0}{Initial prior probability \( p_0 \) for the null distribution in the EM algorithm (default: 0.5).}

\item{tol}{Convergence threshold for the EM algorithm. The algorithm terminates when 
the maximum absolute difference between consecutive gamma values falls below \code{tol} (default: 5e-6).}

\item{is_pvalue}{Logical; if TRUE, the input is assumed to be p-values and transformed accordingly (default: FALSE).}

\item{alternative}{A logical vector indicating whether the alternative distribution is greater (`TRUE`) 
or less (`FALSE`) than the null distribution in each dimension. It must match the number of columns in `z`.}

\item{min_iter}{Minimum number of iterations for the EM algorithm (default: 3).}

\item{max_iter}{Maximum number of iterations for the EM algorithm (default: 50).}

\item{thre_z}{Threshold for gamma values used in log-concave estimation during the M-step of the EM algorithm (default: 0.5).}

\item{Uthre_gam}{Upper threshold for gamma to determine stopping criteria in the EM algorithm (default: 0.99).}

\item{Lthre_gam}{Lower threshold for gamma to determine stopping criteria in the EM algorithm (default: 0.01).}
}
\value{
A list containing estimates from the semiparametric mixture model, including:
  \item{z}{Matrix of input data points.}
  \item{p0}{Estimated prior probability for the null distribution.}
  \item{mu0, sig0}{Parameters of the Gaussian null distribution, \( N(\mu_0, \sigma_0^2) \).}
  \item{f}{Estimated mixture density.}
  \item{f1}{Estimated log-concave alternative density.}
  \item{localFDR}{Estimated local false discovery rates.}
  \item{FDR}{Estimated false discovery rates.}
  \item{iter}{Total number of EM iterations performed.}
  \item{dim}{Dimensionality of the input data.}
  \item{alternative}{Vector indicating whether the alternative distribution is greater (`TRUE`) or less (`FALSE`) than the null distribution.}
}
\description{
The `SPMix` function estimates the local false discovery rate (localFDR) and semiparametric mixture density 
from multi-dimensional inputs such as z-values, p-values, or raw data. It employs a two-component 
semiparametric mixture model, integrating Efron's empirical null principle and log-concave density 
estimation for the alternative distribution.
}
