% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Velocity of Galaxies}
\format{
A data frame with two variables:
\describe{
\item{\code{radius}}{Angular measurement of each globular cluster and star (arcmin)}
\item{\code{velocity}}{Radial velocities (km/s)}
}
}
\usage{
galaxy
}
\description{
Radial velocities of globular clusters of M104 with average v=1121 km/s,
and Milky Way stars which have much lower radial velocities.
}
\keyword{datasets}
