% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{P-values for Pathways.}
\format{
A data frame with four variables:
\describe{
\item{\code{Gene.Set}}{The name of pathways}
\item{\code{L}}{P-values of upregulated genes in peripheral leukocytes}
\item{\code{O}}{P-values of upregulated genes in orbital inflammatory disease}
\item{\code{S}}{P-values of upregulated genes in sinus brushings}
}
}
\usage{
pathways
}
\description{
The list of significantly upregulated genes in three published studies:
(1) peripheral leukocytes(L), (2) orbital inflammatory disease(O), and
(3) sinus brushings(S) compared to healthy controls.
}
\keyword{datasets}
