% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFDR.R
\name{plotFDR}
\alias{plotFDR}
\title{plotFDR estimation for given z-values}
\usage{
plotFDR(z, p0, mu0, sig0, f1, localFDR, leftNull = TRUE, thre_localFDR = 0.2)
}
\arguments{
\item{z}{Matrix which column indicates z-values, probit-transformed p-values.}

\item{p0}{Prior probability for null distribution}

\item{mu0}{sig0 Parameter estimates of normal null distribution, N(mu0, sig0^2)}

\item{f1}{Probability estimates of alternative distribution of mixture model for each z-value point.}

\item{localFDR}{localFDR estimates for given z-values}

\item{leftNull}{If TRUE, a null distribution is placed to the left of the alternative distribution. (default: TRUE)}

\item{thre_localFDR}{Threshold of localFDR for null and alternative distribution (default: 0.2)}
}
\value{
Plot estimated semiparametric mixture density and return threshold value.

  \item{thre}{Threshold z-value for null and alternative distribution}
}
\description{
\code{plotFDR} returns plotFDR estimates and semiparametric
mixture density estimates for given multi-dimensional lists of z-values, which
are the probit-transformed p-values.
For the hypothesis testing \code{plotFDR} uses a two-component semiparametric
mixture model to estimate the plotFDR from the p-values. The two pillars of the
proposed approach are Efron's empirical null principle and log-concave density
estimation for the alternative distribution.
}
