% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPMix.R
\name{plotSPMix}
\alias{plotSPMix}
\title{plotMixture estimation for given z-values}
\usage{
plotSPMix(
  x,
  thre_localFDR = 0.2,
  testing = TRUE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  coord_legend = c(8, -5, 0.2)
)
}
\arguments{
\item{x}{object of class "SPMix"}

\item{thre_localFDR}{Threshold of localFDR determining significant data points. (default: 0.2)}

\item{testing}{If TRUE, it's for hypothesis testing. If FALSE, it's for density estimation (default: TRUE)}

\item{xlab}{Label for x-axis on histogram or 3D scatter plot (default: "x")}

\item{ylab}{Label for y-axis on 3D scatter plot (default: "y")}

\item{zlab}{Label for z-axis on 3D scatter plot (default: "z")}

\item{coord_legend}{Coordinate of a legend for a 3d scatter plot when given data is 2D or 3D. (default: c(8, -5, 0.2))}
}
\value{
Plot estimated semiparametric mixture density and return threshold value.

  \item{thre}{Threshold z-value for null and alternative distribution}
}
\description{
\code{plotSPMix} returns plotSPMix estimates and semiparametric
mixture density estimates for given multi-dimensional lists of z-values, which
are the probit-transformed p-values.
For the hypothesis testing \code{plotSPMix} uses a two-component semiparametric
mixture model to estimate the plotSPMix from the p-values. The two pillars of the
proposed approach are Efron's empirical null principle and log-concave density
estimation for the alternative distribution.
}
