% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPMix.R
\name{plotSPMix}
\alias{plotSPMix}
\title{draws a histogram(univariate) or a scatterplot(2D or 3D) for fitted SPMix object.}
\usage{
plotSPMix(
  x,
  thre_localFDR = 0.2,
  testing = TRUE,
  xlab = "x",
  ylab = "y",
  zlab = "z",
  coord_legend = c(8, -5, 0.2)
)
}
\arguments{
\item{x}{object of class "SPMix"}

\item{thre_localFDR}{Threshold of localFDR determining significant data points. (default: 0.2)}

\item{testing}{If TRUE, it's for hypothesis testing. If FALSE, it's for density estimation (default: TRUE)}

\item{xlab}{Label for x-axis on histogram or 3D scatter plot (default: "x")}

\item{ylab}{Label for y-axis on 3D scatter plot (default: "y")}

\item{zlab}{Label for z-axis on 3D scatter plot (default: "z")}

\item{coord_legend}{Coordinate of a legend for a 3d scatter plot when given data is 2D or 3D. (default: c(8, -5, 0.2))}
}
\value{
Plot estimated semiparametric mixture density and return threshold value.

  \item{thre}{Threshold z-value for null and alternative distribution}
}
\description{
\code{plotSPMix} draws a histogram(univariate) or a scatterplot(2D or 3D) 
for fitted SPMix object which can be customized with the ggplot2 package.
\code{plotSPMix()} can be used for both hypothesis testing and density estimation. 
If testing is TRUE (default), it gives fitted density for both null and alternative distribution.
}
