% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPMix.R
\name{plotSPMix}
\alias{plotSPMix}
\title{plotMixture estimation for given z-values}
\usage{
plotSPMix(
  z,
  p0,
  mu0,
  sig0,
  f,
  f1,
  localFDR,
  p_value = FALSE,
  alternative = "greater",
  thre_localFDR = 0.2,
  testing = TRUE,
  xlab = "",
  ylab = "",
  type = "3d"
)
}
\arguments{
\item{z}{Matrix which column indicates z-values, probit-transformed p-values.}

\item{p0}{Prior probability for null distribution}

\item{mu0}{sig0 Parameter estimates of normal null distribution, N(mu0, sig0^2)}

\item{f1}{Probability estimates of alternative distribution of mixture model for each z-value point.}

\item{localFDR}{localFDR estimates for given z-values}

\item{p_value}{If TRUE, the column of input indicates p-values, if FALSE, it indicates z-values or raw data. (default: FALSE)}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided", "greater" (default) or "less". You can specify just the initial letter. (default: "greater")}

\item{thre_localFDR}{Threshold of localFDR for null and alternative distribution (default: 0.2)}

\item{testing}{TRUE if it's for hypothesis testing, FALSEif it's for density estimation}

\item{type}{Type of 2-dimensional density plot (3d/contour plot)(default: "3d")}
}
\value{
Plot estimated semiparametric mixture density and return threshold value.

  \item{thre}{Threshold z-value for null and alternative distribution}
}
\description{
\code{plotSPMix} returns plotSPMix estimates and semiparametric
mixture density estimates for given multi-dimensional lists of z-values, which
are the probit-transformed p-values.
For the hypothesis testing \code{plotSPMix} uses a two-component semiparametric
mixture model to estimate the plotSPMix from the p-values. The two pillars of the
proposed approach are Efron's empirical null principle and log-concave density
estimation for the alternative distribution.
}
