% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPMix.R
\name{SPMix}
\alias{SPMix}
\title{Semiparametric Mixture Density Estimation for given z-values}
\usage{
SPMix(
  z,
  tol = 5e-06,
  p_value = FALSE,
  alternative = "greater",
  max_iter = 30,
  mono = TRUE,
  thre_z = 0.9,
  Uthre_gam = 0.9,
  Lthre_gam = 0.01
)
}
\arguments{
\item{z}{Matrix which column indicates z-values, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm. If maximum absolute difference
of current and previous gamma value is smaller than tol,
i.e. \eqn{max_i |\gamma_i^{(k+1)}-\gamma_i^{(k)} <tol}, for k-th step,
then optimization stops. (default: 5e-6)}

\item{p_value}{If TRUE, the column of input indicates p-values, if FALSE, it indicates z-values or raw data. (default: FALSE)}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided", "greater" (default) or "less". You can specify just the initial letter. (default: "greater")}

\item{max_iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{mono}{If TRUE, localFDR is in ascending order of z-values. (default: TRUE)}

\item{thre_z}{Threshold value which only z-values smaller than thre.z
are used to compute the log-concave estimates f_1 in M-step.}

\item{Uthre_gam}{Upper threshold of gamma which are used to compute stopping criteria for the EM algorithm.}

\item{Lthre_gam}{Lower threshold of gamma which are used to compute stopping criteria for the EM algorithm.}
}
\value{
Estimates of semiparametric mixture model for f for given z-values.

  \item{p0}{Prior probability for null distribution}
  \item{mu0 sig0}{Parameter estimates of normal null distribution, N(mu0, sig0^2)}
  \item{f}{Probability estimates of semiparametric mixture model for each z-value point.}
  \item{f1}{Probability estimates of alternative distribution of mixture model for each z-value point.}
  \item{localFDR}{localFDR estimates for given z-values}
  \item{iter}{Number of iterations of EM algorithm to compute localFDR.}
}
\description{
\code{SPMix} returns localFDR estimates and semiparametric
mixture density estimates for given multi-dimensional lists of z-values, which
are the probit-transformed p-values.
For the hypothesis testing \code{SPMix} uses a two-component semiparametric
mixture model to estimate the localFDR from the z-values. The two pillars of the
proposed approach are Efron's empirical null principle and log-concave density
estimation for the alternative distribution.
}
